requirements.txt.txt
streamlit
gdown
tensorflow
pillow
numpy
